requirements.txt
# Core
python-dotenv==1.0.0
supabase==2.4.0
anthropic==0.26.0
httpx==0.27.0
pydantic==2.7.0

# Data collection
pytrends==4.9.2
praw==7.7.1
yfinance==0.2.40
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.2.1

# NLP / embeddings
sentence-transformers==3.0.1
scikit-learn==1.4.2
numpy==1.26.4

# Scheduling / pipeline
schedule==1.2.2
tenacity==8.3.0

# API
fastapi==0.111.0
uvicorn==0.30.1

# Utils
loguru==0.7.2
python-dateutil==2.9.0
